chore(web): remove jobs page
Delete the jobs routes, content, and related web references.
