+++ apps/web/content-collections.ts
@@ -759,46 +759,6 @@ const updates = defineCollection({
   },
 });
 
-const jobs = defineCollection({
-  name: "jobs",
-  directory: "content/jobs",
-  include: "*.mdx",
-  exclude: "AGENTS.md",
-  schema: z.object({
-    title: z.string(),
-    description: z.string(),
-    cardDescription: z.string(),
-    backgroundImage: z.string(),
-    applyUrl: z.string().optional(),
-    published: z.boolean().default(true),
-  }),
-  transform: async (document, context) => {
-    const mdx = await compileMDX(context, document, {
-      remarkPlugins: [remarkGfm, mdxMermaid],
-      rehypePlugins: [
-        rehypeSlug,
-        [
-          rehypeAutolinkHeadings,
-          {
-            behavior: "wrap",
-            properties: {
-              className: ["anchor"],
-            },
-          },
-        ],
-      ],
-    });
-
-    const slug = document._meta.path.replace(/\.mdx$/, "");
-
-    return {
-      ...document,
-      mdx,
-      slug,
-    };
-  },
-});
-
 const solutions = defineCollection({
   name: "solutions",
   directory: "content/solutions",
@@ -868,7 +828,6 @@ export default defineConfig({
     handbook,
     roadmap,
     ossFriends,
-    jobs,
     updates,
     solutions,
   ],
