--- apps/web/content/jobs/engineer.mdx
@@ -1,56 +1,0 @@
----
-published: false
-title: "Engineer"
-description: "Join Char as a Product-Minded Engineer. Build with Tauri, Rust, and TypeScript."
-cardDescription: "Build great software with passion. Work on our desktop app, web platform, and AI features. Join a team that cares about craft and quality."
-backgroundImage: "/api/images/beach.png"
-applyUrl: "mailto:yujonglee@char.com?cc=john@char.com&subject=Application for Engineer"
----
-
-This is written by <GitHubMention username="yujonglee" name="Yujong" />, who you'll be directly working with.
-
-## Why are we hiring?
-
-Char is product-led. Engineering execution is our bottleneck.
-We need someone to own responsibility for a certain part of the codebase and act as a first-line reviewer.
-
-<Collapsible title="And we're doing pretty well these days">
-<CompanyStats />
-</Collapsible>
-
-## What are we expecting?
-
-We need someone who:
-
-- Can take full responsibility for a certain part of the codebase.
-- Can define a problem & a solution without relying on others.
-
-## How to apply?
-
-1. Read the code and browse existing issues in the [repository](https://github.com/fastrepl/char).
-2. Feel free to tag <GitHubMention username="yujonglee" name="Yujong" /> if you want to discuss something.
-3. Propose a problem/solution through an issue, comment, or pull request.
-4. [Send us an email](mailto:founders@char.com) that references step 3. We are not really looking for a resume. Something like GH account where I can see your past work and interests would be enough.
-
-## Who will be a good fit?
-
-We are not judging anything, but we just want to find the best fit for the role.
-
-<Collapsible title="Click to read more">
-You're likely aligned with us if:
-
-- You spend most of your time thinking, refactoring, and designing better systems, and leave most of the implementation to AI.
-- You love documentation and writing. You are happy to do work in an async manner.
-
-You are not likely aligned with us if:
-
-- You are obsessed with Claude code hype like Ralph loop
-- You are optimized for token consumption and code generation.
-- You are looking for other people to teach you and tell you what to do.
-</Collapsible>
-
-## What to expect?
-
-- Remote-first. Async via <ToolIcon icon="logos:slack-icon" /> Slack. Occasional calls with <CharIcon /> Char on.
-- You write open source code every day. Everything runs through <ToolIcon icon="logos:github-icon" /> GitHub.
-- `$80 ~ 150k/year`, `0.5 ~ 1.5%` equity. Depends on scope and experience.
